+++ templates/settings.php
@@ -48,7 +48,11 @@
         </p>
 
         <p class="onlyoffice-header"><?php p($l->t("Secret key (leave blank to disable)")) ?></p>
-        <p><input id="onlyofficeSecret" value="<?php p($_["secret"]) ?>" placeholder="secret" type="text"></p>
+        <p>
+            <input id="onlyofficeSecret" value="<?php p($_["secret"]) ?>" placeholder="secret" type="password" />
+            <input type="checkbox" id="personal-show" name="show">
+            <label id="onlyofficeSecret-show" for="personal-show"></label>
+        </p>
 
         <p>
             <a id="onlyofficeAdv" class="onlyoffice-header">
@@ -100,7 +104,7 @@
         <p>
             <input type="checkbox" class="checkbox" id="onlyofficePreview"
                 <?php if ($_["preview"]) { ?>checked="checked"<?php } ?> />
-            <label for="onlyofficePreview"><?php p($l->t("Generate document preview")) ?> <?php p($l->t("(take up disk space)")) ?></label>
+            <label for="onlyofficePreview"><?php p($l->t("Use ONLYOFFICE to generate a document preview (it will take up disk space)")) ?></label>
         </p>
 
         <p>