+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## Added
 - viewing a locked file
+- hide secret key in settings
 - configuring version storage
 - clearing history data
+
+## Changed
+- thumbnails for small files only
 
 ## 6.2.1
 ## Changed